site/requirements.txt
azure-identity
azure-keyvault-secrets
redis
pycentral
ipapi
loguru